requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
gunicorn==21.2.0
python-dateutil==2.8.2
requests==2.31.0
email_validator
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
python-magic==0.4.27
Flask-Limiter==3.5.0
